+++ app/containers/LoginServices.js
@@ -1,0 +1,370 @@
+import React from 'react';
+import {
+	View, StyleSheet, Text, Animated, Easing, Image
+} from 'react-native';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { Base64 } from 'js-base64';
+import { withNavigation } from 'react-navigation';
+
+import { withTheme } from '../theme';
+import sharedStyles from '../views/Styles';
+import { themes } from '../constants/colors';
+import { loginRequest as loginRequestAction } from '../actions/login';
+import Button from './Button';
+import OnboardingSeparator from './OnboardingSeparator';
+import Touch from '../utils/touch';
+import i18n from '../i18n';
+import random from '../utils/random';
+
+const SERVICE_HEIGHT = 58;
+const SERVICES_COLLAPSED_HEIGHT = 174;
+
+const styles = StyleSheet.create({
+	serviceButton: {
+		borderRadius: 2,
+		marginBottom: 10
+	},
+	serviceButtonContainer: {
+		borderRadius: 2,
+		width: '100%',
+		height: 48,
+		flexDirection: 'row',
+		alignItems: 'center',
+		justifyContent: 'center',
+		paddingHorizontal: 15
+	},
+	serviceIcon: {
+		position: 'absolute',
+		left: 15,
+		top: 12,
+		width: 24,
+		height: 24
+	},
+	serviceText: {
+		...sharedStyles.textRegular,
+		fontSize: 16
+	},
+	serviceName: {
+		...sharedStyles.textSemibold
+	}
+});
+
+class LoginServices extends React.PureComponent {
+	static propTypes = {
+		navigation: PropTypes.object,
+		server: PropTypes.string,
+		services: PropTypes.object,
+		Gitlab_URL: PropTypes.string,
+		CAS_enabled: PropTypes.bool,
+		CAS_login_url: PropTypes.string,
+		Accounts_ShowFormLogin: PropTypes.bool,
+		Accounts_RegistrationForm: PropTypes.string,
+		theme: PropTypes.string
+	}
+
+	state = {
+		collapsed: true,
+		servicesHeight: new Animated.Value(SERVICES_COLLAPSED_HEIGHT),
+	}
+
+	onPressFacebook = () => {
+		const { services, server } = this.props;
+		const { clientId } = services.facebook;
+		const endpoint = 'https://m.facebook.com/v2.9/dialog/oauth';
+		const redirect_uri = `${ server }/_oauth/facebook?close`;
+		const scope = 'email';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }&display=touch`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressGithub = () => {
+		const { services, server } = this.props;
+		const { clientId } = services.github;
+		const endpoint = `https://github.com/login?client_id=${ clientId }&return_to=${ encodeURIComponent('/login/oauth/authorize') }`;
+		const redirect_uri = `${ server }/_oauth/github?close`;
+		const scope = 'user:email';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }`;
+		this.openOAuth({ url: `${ endpoint }${ encodeURIComponent(params) }` });
+	}
+
+	onPressGitlab = () => {
+		const { services, server, Gitlab_URL } = this.props;
+		const { clientId } = services.gitlab;
+		const baseURL = Gitlab_URL ? Gitlab_URL.trim().replace(/\/*$/, '') : 'https://gitlab.com';
+		const endpoint = `${ baseURL }/oauth/authorize`;
+		const redirect_uri = `${ server }/_oauth/gitlab?close`;
+		const scope = 'read_user';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }&response_type=code`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressGoogle = () => {
+		const { services, server } = this.props;
+		const { clientId } = services.google;
+		const endpoint = 'https://accounts.google.com/o/oauth2/auth';
+		const redirect_uri = `${ server }/_oauth/google?close`;
+		const scope = 'email';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }&response_type=code`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressLinkedin = () => {
+		const { services, server } = this.props;
+		const { clientId } = services.linkedin;
+		const endpoint = 'https://www.linkedin.com/oauth/v2/authorization';
+		const redirect_uri = `${ server }/_oauth/linkedin?close`;
+		const scope = 'r_liteprofile,r_emailaddress';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }&response_type=code`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressMeteor = () => {
+		const { services, server } = this.props;
+		const { clientId } = services['meteor-developer'];
+		const endpoint = 'https://www.meteor.com/oauth2/authorize';
+		const redirect_uri = `${ server }/_oauth/meteor-developer`;
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&state=${ state }&response_type=code`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressTwitter = () => {
+		const { server } = this.props;
+		const state = this.getOAuthState();
+		const url = `${ server }/_oauth/twitter/?requestTokenAndRedirect=true&state=${ state }`;
+		this.openOAuth({ url });
+	}
+
+	onPressWordpress = () => {
+		const { services, server } = this.props;
+		const { clientId, serverURL } = services.wordpress;
+		const endpoint = `${ serverURL }/oauth/authorize`;
+		const redirect_uri = `${ server }/_oauth/wordpress?close`;
+		const scope = 'openid';
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirect_uri }&scope=${ scope }&state=${ state }&response_type=code`;
+		this.openOAuth({ url: `${ endpoint }${ params }` });
+	}
+
+	onPressCustomOAuth = (loginService) => {
+		const { server } = this.props;
+		const {
+			serverURL, authorizePath, clientId, scope, service
+		} = loginService;
+		const redirectUri = `${ server }/_oauth/${ service }`;
+		const state = this.getOAuthState();
+		const params = `?client_id=${ clientId }&redirect_uri=${ redirectUri }&response_type=code&state=${ state }&scope=${ scope }`;
+		const domain = `${ serverURL }`;
+		const absolutePath = `${ authorizePath }${ params }`;
+		const url = absolutePath.includes(domain) ? absolutePath : domain + absolutePath;
+		this.openOAuth({ url });
+	}
+
+	onPressSaml = (loginService) => {
+		const { server } = this.props;
+		const {	clientConfig } = loginService;
+		const {	provider } = clientConfig;
+		const ssoToken = random(17);
+		const url = `${ server }/_saml/authorize/${ provider }/${ ssoToken }`;
+		this.openOAuth({ url, ssoToken, authType: 'saml' });
+	}
+
+	onPressCas = () => {
+		const { server, CAS_login_url } = this.props;
+		const ssoToken = random(17);
+		const url = `${ CAS_login_url }?service=${ server }/_cas/${ ssoToken }`;
+		this.openOAuth({ url, ssoToken, authType: 'cas' });
+	}
+
+	getOAuthState = () => {
+		const credentialToken = random(43);
+		return Base64.encodeURI(JSON.stringify({ loginStyle: 'popup', credentialToken, isCordova: true }));
+	}
+
+	openOAuth = ({ url, ssoToken, authType = 'oauth' }) => {
+		const { navigation } = this.props;
+		navigation.navigate('AuthenticationWebView', { url, authType, ssoToken });
+	}
+
+	transitionServicesTo = (height) => {
+		const { servicesHeight } = this.state;
+		if (this._animation) {
+			this._animation.stop();
+		}
+		this._animation = Animated.timing(servicesHeight, {
+			toValue: height,
+			duration: 300,
+			easing: Easing.easeOutCubic
+		}).start();
+	}
+
+	toggleServices = () => {
+		const { collapsed } = this.state;
+		const { services } = this.props;
+		const { length } = Object.values(services);
+		if (collapsed) {
+			this.transitionServicesTo(SERVICE_HEIGHT * length);
+		} else {
+			this.transitionServicesTo(SERVICES_COLLAPSED_HEIGHT);
+		}
+		this.setState(prevState => ({ collapsed: !prevState.collapsed }));
+	}
+
+	getSocialOauthProvider = (name) => {
+		const oauthProviders = {
+			facebook: this.onPressFacebook,
+			github: this.onPressGithub,
+			gitlab: this.onPressGitlab,
+			google: this.onPressGoogle,
+			linkedin: this.onPressLinkedin,
+			'meteor-developer': this.onPressMeteor,
+			twitter: this.onPressTwitter,
+			wordpress: this.onPressWordpress
+		};
+		return oauthProviders[name];
+	}
+
+	renderServicesSeparator = () => {
+		const { collapsed } = this.state;
+		const {
+			services, theme, Accounts_ShowFormLogin, Accounts_RegistrationForm
+		} = this.props;
+		const { length } = Object.values(services);
+
+		if (length > 3 && Accounts_ShowFormLogin && Accounts_RegistrationForm) {
+			return (
+				<>
+					<Button
+						title={collapsed ? 'More options' : 'Less options'}
+						type='secondary'
+						onPress={this.toggleServices}
+						theme={theme}
+						style={{ marginBottom: 0 }}
+						color={themes[theme].actionTintColor}
+					/>
+					<OnboardingSeparator theme={theme} />
+				</>
+			);
+		}
+		if (length > 0) {
+			return <OnboardingSeparator theme={theme} />;
+		}
+		return null;
+	}
+
+	renderItem = (service) => {
+		let { name } = service;
+		name = name === 'meteor-developer' ? 'meteor' : name;
+		const icon = `icon_${ name }`;
+		const isSaml = service.service === 'saml';
+		let onPress = () => {};
+
+		switch (service.authType) {
+			case 'oauth': {
+				onPress = this.getSocialOauthProvider(service.name);
+				break;
+			}
+			case 'oauth_custom': {
+				onPress = () => this.onPressCustomOAuth(service);
+				break;
+			}
+			case 'saml': {
+				onPress = () => this.onPressSaml(service);
+				break;
+			}
+			case 'cas': {
+				onPress = () => this.onPressCas();
+				break;
+			}
+			default:
+				break;
+		}
+		name = name.charAt(0).toUpperCase() + name.slice(1);
+		const { CAS_enabled, theme } = this.props;
+		let buttonText;
+		if (isSaml || (service.service === 'cas' && CAS_enabled)) {
+			buttonText = <Text style={[styles.serviceName, isSaml && { color: service.buttonLabelColor }]}>{name}</Text>;
+		} else {
+			buttonText = (
+				<>
+					{i18n.t('Continue_with')} <Text style={styles.serviceName}>{name}</Text>
+				</>
+			);
+		}
+
+		const backgroundColor = isSaml && service.buttonColor ? service.buttonColor : themes[theme].chatComponentBackground;
+
+		return (
+			<Touch
+				key={service.name}
+				onPress={onPress}
+				style={[styles.serviceButton, { backgroundColor }]}
+				theme={theme}
+				activeOpacity={0.5}
+				underlayColor={themes[theme].buttonText}
+			>
+				<View style={styles.serviceButtonContainer}>
+					{service.authType === 'oauth' ? <Image source={{ uri: icon }} style={styles.serviceIcon} /> : null}
+					<Text style={[styles.serviceText, { color: themes[theme].titleText }]}>{buttonText}</Text>
+				</View>
+			</Touch>
+		);
+	}
+
+	render() {
+		const { servicesHeight } = this.state;
+		const { services, Accounts_ShowFormLogin, Accounts_RegistrationForm } = this.props;
+		const { length } = Object.values(services);
+		const style = {
+			overflow: 'hidden',
+			height: servicesHeight
+		};
+
+		if (length > 3 && Accounts_ShowFormLogin && Accounts_RegistrationForm) {
+			return (
+				<>
+					<Animated.View style={style}>
+						{Object.values(services).map(service => this.renderItem(service))}
+					</Animated.View>
+					{this.renderServicesSeparator()}
+				</>
+			);
+		}
+		return (
+			<>
+				{Object.values(services).map(service => this.renderItem(service))}
+				{this.renderServicesSeparator()}
+			</>
+		);
+	}
+}
+
+const mapStateToProps = state => ({
+	server: state.server.server,
+	Site_Name: state.settings.Site_Name,
+	Gitlab_URL: state.settings.API_Gitlab_URL,
+	CAS_enabled: state.settings.CAS_enabled,
+	CAS_login_url: state.settings.CAS_login_url,
+	Accounts_ShowFormLogin: state.settings.Accounts_ShowFormLogin,
+	Accounts_RegistrationForm: state.settings.Accounts_RegistrationForm,
+	Accounts_RegistrationForm_LinkReplacementText: state.settings.Accounts_RegistrationForm_LinkReplacementText,
+	services: state.login.services,
+	isFetching: state.login.isFetching,
+	failure: state.login.failure,
+	error: state.login.error && state.login.error.data,
+	Accounts_EmailOrUsernamePlaceholder: state.settings.Accounts_EmailOrUsernamePlaceholder,
+	Accounts_PasswordPlaceholder: state.settings.Accounts_PasswordPlaceholder,
+	Accounts_PasswordReset: state.settings.Accounts_PasswordReset
+});
+
+const mapDispatchToProps = dispatch => ({
+	loginRequest: params => dispatch(loginRequestAction(params))
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(withTheme(withNavigation(LoginServices)));
